Fixed issues with screen capture functionality
